Update Volatility feature to use new Feature base class - fix Feature._rolling_std and adjust market_data.Scalar to ensure tests pass

+++ alpha_research_framework/market_data/scalar.py
@@ -2,5 +2,5 @@ from typing import TypeAlias
 
 import numpy as np
 
-Scalar: TypeAlias = np.float32
+Scalar: TypeAlias = np.float64
 """Datatype used in all market data arrays."""
